Add closed property useful to heartbeaters
This property allows heartbeaters to avoid sending HEARTBEAT command while a closing handshake is in progress.

This was an issue previously, as the CloseDiscordConnection exception is not raised until the closing handshake is complete or requires a response.

This means that a heartbeat could be sent between the client initiating the handshake and receiving the response.